ubiquity-frontend-max (7.0.max16)   * Don't kill pluginin ubiquity-frontend-max (7.0.max15)   * Enable prepare   * Move intro to usersetup step ubiquity-frontend-max (7.0.max14)   * Enable prepare module
git-svn-id: http://max.educa.madrid.org:8000/svn/trunk@1178 424cea3e-0240-0410-9575-8bd47b58cd33

+++ ubiquity-frontend-max/usr/share/ubiquity/plugininstall.py
@@ -1755,6 +1755,7 @@ class Install(install_misc.InstallBase):
                 if pids[pid][1].startswith("python"): continue
                 if pids[pid][1].startswith("install"): continue
                 if pids[pid][1].startswith("plugininstall"): continue
+                if pids[pid][1].startswith("pluginin"): continue
                 syslog.syslog("DEBUG: killing %s" %(pids[pid]) )
                 pin,pout=os.popen2("sudo kill %s"%(pid) )
                 pin.close()
